requirements.txt
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-gis==1.0
drf-yasg==1.21.4
environs==9.5.0
model-bakery==1.9.0
psycopg2-binary==2.9.5
